+++ app/dashboard/page.tsx
@@ -1,56 +1,33 @@
 "use client";
 
 import Navbar from "@/components/navbar";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import {
-  Dialog,
-  DialogHeader,
-  DialogContent,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
 import { useBoards } from "@/lib/hooks/useBoards";
 import { Board } from "@/lib/supabase/models";
-import { useUser } from "@clerk/nextjs";
-import {
-  Filter,
-  Grid3x3,
-  List,
-  Plus,
-  Rocket,
-  Search,
-  Trello,
-} from "lucide-react";
-import Link from "next/link";
-import { useRouter } from "next/navigation";
+import { BoardFilterData } from "@/lib/types";
 import { useState } from "react";
+import {
+  DashboardHeader,
+  DashboardStats,
+  BoardsSection,
+  FilterDialog,
+  UpgradeDialog,
+} from "@/components/dashboard";
 
 export default function DashboardPage() {
-  const { user } = useUser();
   const { createBoard, boards, error } = useBoards();
-  const router = useRouter();
   const [viewMode, setViewMode] = useState<"grid" | "list">("grid");
   const [isFilterOpen, setIsFilterOpen] = useState<boolean>(false);
   const [showUpgradeDialog, setShowUpgradeDialog] = useState<boolean>(false);
 
-  const [filters, setFilters] = useState({
+  const [filters, setFilters] = useState<BoardFilterData>({
     search: "",
     dateRange: {
-      start: null as string | null,
-      end: null as string | null,
+      start: null,
+      end: null,
     },
     taskCount: {
-      min: null as number | null,
-      max: null as number | null,
+      min: null,
+      max: null,
     },
   });
 
@@ -77,12 +54,12 @@ export default function DashboardPage() {
     setFilters({
       search: "",
       dateRange: {
-        start: null as string | null,
-        end: null as string | null,
+        start: null,
+        end: null,
       },
       taskCount: {
-        min: null as number | null,
-        max: null as number | null,
+        min: null,
+        max: null,
       },
     });
   }
@@ -105,375 +82,34 @@ export default function DashboardPage() {
       <Navbar />
 
       <main className="container mx-auto px-4 py-6 sm:py-8">
-        <div className="mb-6 sm:mb-8">
-          <h1 className="text-2xl sm:text-3xl font-bold text-gray-900 mb-2">
-            Welcome back,{" "}
-            {user?.firstName ?? user?.emailAddresses[0].emailAddress}!
-          </h1>
-          <p className="text-gray-600">
-            Here what happening with your boards today.
-          </p>
-        </div>
-
-        {/* Stats */}
-        <div className="grid grid-cols-2 lg:grid-cols-4 gap-4 sm:gap-6 mb-6 sm:mb-8">
-          <Card>
-            <CardContent className="p-4 sm:p-6">
-              <div className="flex items-center justify-between">
-                <div>
-                  <p className="text-xs sm:text-sm font-medium text-gray-600">
-                    Total Boards
-                  </p>
-                  <p className="text-xl sm:text-2xl font-bold text-gray-900">
-                    {boards.length}
-                  </p>
-                </div>
-                <div className="h-10 w-10 sm:h-12 sm:w-12 bg-blue-100 rounded-lg flex items-center justify-center">
-                  <Trello className="h-5 w-5 sm:h-6 sm:w-6 text-blue-600" />
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-          <Card>
-            <CardContent className="p-4 sm:p-6">
-              <div className="flex items-center justify-between">
-                <div>
-                  <p className="text-xs sm:text-sm font-medium text-gray-600">
-                    Active Projects
-                  </p>
-                  <p className="text-xl sm:text-2xl font-bold text-gray-900">
-                    {boards.length}
-                  </p>
-                </div>
-                <div className="h-10 w-10 sm:h-12 sm:w-12 bg-green-100 rounded-lg flex items-center justify-center">
-                  <Rocket className="h-5 w-5 sm:h-6 sm:w-6 text-green-600" />
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-          <Card>
-            <CardContent className="p-4 sm:p-6">
-              <div className="flex items-center justify-between">
-                <div>
-                  <p className="text-xs sm:text-sm font-medium text-gray-600">
-                    Recent Activity
-                  </p>
-                  <p className="text-xl sm:text-2xl font-bold text-gray-900">
-                    {
-                      boards.filter((board) => {
-                        const updatedAt = new Date(board.updated_at);
-                        const oneWeekAgo = new Date();
-                        oneWeekAgo.setDate(oneWeekAgo.getDate() - 7);
-                        return updatedAt > oneWeekAgo;
-                      }).length
-                    }
-                  </p>
-                </div>
-                <div className="h-10 w-10 sm:h-12 sm:w-12 bg-purple-100 rounded-lg flex items-center justify-center">
-                  📊
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-          <Card>
-            <CardContent className="p-4 sm:p-6">
-              <div className="flex items-center justify-between">
-                <div>
-                  <p className="text-xs sm:text-sm font-medium text-gray-600">
-                    Total Boards
-                  </p>
-                  <p className="text-xl sm:text-2xl font-bold text-gray-900">
-                    {boards.length}
-                  </p>
-                </div>
-                <div className="h-10 w-10 sm:h-12 sm:w-12 bg-blue-100 rounded-lg flex items-center justify-center">
-                  <Trello className="h-5 w-5 sm:h-6 sm:w-6 text-blue-600" />
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-        </div>
-
-        {/* Boards */}
-        <div className="mb-6 sm:mb-8">
-          <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between mb-4 sm:mb-6 space-y-4 sm:space-y-0">
-            <div>
-              <h2 className="text-xl sm:text-2xl font-bold text-gray-900">
-                Your Boards
-              </h2>
-              <p className="text-gray-600">Manage your projects and tasks</p>
-            </div>
-
-            <div className="flex flex-col sm:flex-row items-stretch sm:items-center space-y-2 sm:space-y-0 sm:space-x-4">
-              <div className="flex items-center space-x-2 rounded bg-white border p-1">
-                <Button
-                  variant={viewMode === "grid" ? "default" : "ghost"}
-                  size="sm"
-                  onClick={() => setViewMode("grid")}
-                >
-                  <Grid3x3 />
-                </Button>
-                <Button
-                  variant={viewMode === "list" ? "default" : "ghost"}
-                  size="sm"
-                  onClick={() => setViewMode("list")}
-                >
-                  <List />
-                </Button>
-              </div>
-
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={() => setIsFilterOpen(true)}
-              >
-                <Filter />
-                Filter
-              </Button>
-
-              <Button onClick={handleCreateBoard}>
-                <Plus />
-                Create Board
-              </Button>
-            </div>
-          </div>
-          {/* Search Bar */}
-          <div className="relative mb-4 sm:mb-6">
-            <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 h-4 w-4 text-gray-400" />
-            <Input
-              id="search"
-              placeholder="Search boards..."
-              className="pl-10"
-              onChange={(e) =>
-                setFilters((prev) => ({ ...prev, search: e.target.value }))
-              }
-            />
-          </div>
-
-          {/* Boards Grid/List */}
-          {boards.length === 0 ? (
-            <div>No boards yet</div>
-          ) : viewMode === "grid" ? (
-            <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4 sm:gap-6">
-              {filteredBoards.map((board, key) => (
-                <Link href={`/boards/${board.id}`} key={key}>
-                  <Card className="hover:shadow-lg transition-shadow cursor-pointer group">
-                    <CardHeader className="pb-3">
-                      <div className="flex items-center justify-between">
-                        <div className={`w-4 h-4 ${board.color} rounded`} />
-                        <Badge className="text-xs" variant="secondary">
-                          New
-                        </Badge>
-                      </div>
-                    </CardHeader>
-                    <CardContent className="p-4 sm:p-6">
-                      <CardTitle className="text-base sm:text-lg mb-2 group-hover:text-blue-600 transition-colors">
-                        {board.title}
-                      </CardTitle>
-                      <CardDescription className="text-sm mb-4">
-                        {board.description}
-                      </CardDescription>
-                      <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between text-xs text-gray-500 space-y-1 sm:space-y-0">
-                        <span>
-                          Created{" "}
-                          {new Date(board.created_at).toLocaleDateString()}
-                        </span>
-                        <span>
-                          Updated{" "}
-                          {new Date(board.updated_at).toLocaleDateString()}
-                        </span>
-                      </div>
-                    </CardContent>
-                  </Card>
-                </Link>
-              ))}
-
-              {/* <Card className="border-2 border-dashed border-gray-300 hover:border-blue-400 transition-colors cursor-pointer group">
-                <CardContent className="p-4 sm:p-6 flex flex-col items-center justify-center h-full min-h-[200px]">
-                  <Plus className="h-6 w-6 sm:h-8 sm:w-8 text-gray-400 group-hover:text-blue-600 mb-2" />
-                  <p className="text-sm sm:text-base text-gray-600 group-hover:text-blue-600 font-medium">
-                    Create new board
-                  </p>
-                </CardContent>
-              </Card> */}
-            </div>
-          ) : (
-            <div>
-              {boards.map((board, key) => (
-                <div key={key} className={key > 0 ? "mt-4" : ""}>
-                  <Link href={`/boards/${board.id}`}>
-                    <Card className="hover:shadow-lg transition-shadow cursor-pointer group">
-                      <CardHeader className="pb-3">
-                        <div className="flex items-center justify-between">
-                          <div className={`w-4 h-4 ${board.color} rounded`} />
-                          <Badge className="text-xs" variant="secondary">
-                            New
-                          </Badge>
-                        </div>
-                      </CardHeader>
-                      <CardContent className="p-4 sm:p-6">
-                        <CardTitle className="text-base sm:text-lg mb-2 group-hover:text-blue-600 transition-colors">
-                          {board.title}
-                        </CardTitle>
-                        <CardDescription className="text-sm mb-4">
-                          {board.description}
-                        </CardDescription>
-                        <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between text-xs text-gray-500 space-y-1 sm:space-y-0">
-                          <span>
-                            Created{" "}
-                            {new Date(board.created_at).toLocaleDateString()}
-                          </span>
-                          <span>
-                            Updated{" "}
-                            {new Date(board.updated_at).toLocaleDateString()}
-                          </span>
-                        </div>
-                      </CardContent>
-                    </Card>
-                  </Link>
-                </div>
-              ))}
-
-              {/* <Card className="mt-4 border-2 border-dashed border-gray-300 hover:border-blue-400 transition-colors cursor-pointer group">
-                <CardContent className="p-4 sm:p-6 flex flex-col items-center justify-center h-full min-h-[200px]">
-                  <Plus className="h-6 w-6 sm:h-8 sm:w-8 text-gray-400 group-hover:text-blue-600 mb-2" />
-                  <p className="text-sm sm:text-base text-gray-600 group-hover:text-blue-600 font-medium">
-                    Create new board
-                  </p>
-                </CardContent>
-              </Card> */}
-            </div>
-          )}
-        </div>
+        <DashboardHeader />
+        <DashboardStats boards={boards} />
+        
+        <BoardsSection
+          boards={filteredBoards}
+          viewMode={viewMode}
+          onViewModeChange={setViewMode}
+          searchValue={filters.search}
+          onSearchChange={(value) =>
+            setFilters((prev) => ({ ...prev, search: value }))
+          }
+          onFilterClick={() => setIsFilterOpen(true)}
+          onCreateBoard={handleCreateBoard}
+        />
       </main>
 
-      {/* Filter Dialog */}
-      <Dialog open={isFilterOpen} onOpenChange={setIsFilterOpen}>
-        <DialogContent className="w-[95vw] max-w-[425px] mx-auto">
-          <DialogHeader>
-            <DialogTitle>Filter Boards</DialogTitle>
-            <p className="text-sm text-gray-600">
-              Filter boards by title, date, or task count.
-            </p>
-          </DialogHeader>
-          <div className="space-y-4">
-            <div className="space-y-2">
-              <Label>Search</Label>
-              <Input
-                id="search"
-                placeholder="Search board titles..."
-                onChange={(e) =>
-                  setFilters((prev) => ({ ...prev, search: e.target.value }))
-                }
-              />
-            </div>
-            <div className="space-y-2">
-              <Label>Date Range</Label>
-              <div className="grid grid-cols-1 sm:grid-cols-2 gap-2">
-                <div>
-                  <Label className="text-xs">Start Date</Label>
-                  <Input
-                    type="date"
-                    onChange={(e) =>
-                      setFilters((prev) => ({
-                        ...prev,
-                        dateRange: {
-                          ...prev.dateRange,
-                          start: e.target.value || null,
-                        },
-                      }))
-                    }
-                  />
-                </div>
-                <div>
-                  <Label className="text-xs">End Date</Label>
-                  <Input
-                    type="date"
-                    onChange={(e) =>
-                      setFilters((prev) => ({
-                        ...prev,
-                        dateRange: {
-                          ...prev.dateRange,
-                          end: e.target.value || null,
-                        },
-                      }))
-                    }
-                  />
-                </div>
-              </div>
-            </div>
-            <div className="space-y-2">
-              <Label>Task Count</Label>
-              <div className="grid grid-cols-1 sm:grid-cols-2 gap-2">
-                <div>
-                  <Label className="text-xs">Minimum</Label>
-                  <Input
-                    type="number"
-                    min="0"
-                    placeholder="Min tasks"
-                    onChange={(e) =>
-                      setFilters((prev) => ({
-                        ...prev,
-                        taskCount: {
-                          ...prev.taskCount,
-                          min: e.target.value ? Number(e.target.value) : null,
-                        },
-                      }))
-                    }
-                  />
-                </div>
-                <div>
-                  <Label className="text-xs">Maximum</Label>
-                  <Input
-                    type="number"
-                    min="0"
-                    placeholder="Max tasks"
-                    onChange={(e) =>
-                      setFilters((prev) => ({
-                        ...prev,
-                        taskCount: {
-                          ...prev.taskCount,
-                          max: e.target.value ? Number(e.target.value) : null,
-                        },
-                      }))
-                    }
-                  />
-                </div>
-              </div>
-            </div>
+      <FilterDialog
+        isOpen={isFilterOpen}
+        onOpenChange={setIsFilterOpen}
+        filters={filters}
+        onFilterChange={setFilters}
+        onClearFilters={clearFilters}
+      />
 
-            <div className="flex flex-col sm:flex-row justify-between pt-4 space-y-2 sm:space-y-0 sm:space-x-2">
-              <Button variant="outline" onClick={clearFilters}>
-                Clear Filters
-              </Button>
-              <Button onClick={() => setIsFilterOpen(false)}>
-                Apply Filters
-              </Button>
-            </div>
-          </div>
-        </DialogContent>
-      </Dialog>
-
-      <Dialog open={showUpgradeDialog} onOpenChange={setShowUpgradeDialog}>
-        <DialogContent className="w-[95vw] max-w-[425px] mx-auto">
-          <DialogHeader>
-            <DialogTitle>Upgrade to Create More Boards</DialogTitle>
-            <p className="text-sm text-gray-600">
-              Free users can only create one board. Upgrade to Pro or Enterprise
-              to create unlimited boards.
-            </p>
-          </DialogHeader>
-          <div className="flex justify-end space-x-4 pt-4">
-            <Button
-              variant="outline"
-              onClick={() => setShowUpgradeDialog(false)}
-            >
-              Cancel
-            </Button>
-            <Button onClick={() => router.push("/pricing")}>View Plans</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
+      <UpgradeDialog
+        isOpen={showUpgradeDialog}
+        onOpenChange={setShowUpgradeDialog}
+      />
     </div>
   );
 }
