+++ components/board/Column.tsx
@@ -1,0 +1,176 @@
+"use client";
+
+import { Badge } from "@/components/ui/badge";
+import { Button } from "@/components/ui/button";
+import {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+import { Textarea } from "@/components/ui/textarea";
+import { ColumnWithTasks } from "@/lib/supabase/models";
+import { TaskData } from "@/lib/types";
+import { MoreHorizontal } from "lucide-react";
+import { useDroppable } from "@dnd-kit/core";
+import {
+  SortableContext,
+  verticalListSortingStrategy,
+} from "@dnd-kit/sortable";
+import TaskCard from "./TaskCard";
+
+interface ColumnProps {
+  column: ColumnWithTasks;
+  onCreateTask: (taskData: TaskData) => Promise<void>;
+  onEditColumn: (column: ColumnWithTasks) => void;
+}
+
+export default function Column({
+  column,
+  onCreateTask,
+  onEditColumn,
+}: ColumnProps) {
+  const { setNodeRef, isOver } = useDroppable({ id: column.id });
+
+  return (
+    <div
+      ref={setNodeRef}
+      className={`w-full lg:flex-shrink-0 lg:w-80 ${
+        isOver ? "bg-blue-50 rounded-lg" : ""
+      }`}
+    >
+      <div
+        className={`bg-white rounded-lg shadow-sm border ${
+          isOver ? "ring-2 ring-blue-300" : ""
+        }`}
+      >
+        {/* Column Header */}
+        <div className="p-3 sm:p-4 border-b">
+          <div className="flex items-center justify-between">
+            <div className="flex items-center space-x-2 min-w-0">
+              <h3 className="font-semibold text-gray-900 text-sm sm:text-base truncate">
+                {column.title}
+              </h3>
+              <Badge variant="secondary" className="text-xs flex-shrink-0">
+                {column.tasks.length}
+              </Badge>
+            </div>
+            <Button
+              variant="ghost"
+              size="sm"
+              className="flex-shrink-0"
+              onClick={() => onEditColumn(column)}
+            >
+              <MoreHorizontal />
+            </Button>
+          </div>
+        </div>
+
+        {/* column content */}
+        <div className="p-2">
+          <SortableContext
+            items={column.tasks.map((task) => task.id)}
+            strategy={verticalListSortingStrategy}
+          >
+            <div className="space-y-3">
+              {column.tasks.map((task, key) => (
+                <TaskCard task={task} key={key} />
+              ))}
+            </div>
+          </SortableContext>
+
+          <Dialog>
+            {/* <DialogTrigger asChild>
+              <Button
+                variant="ghost"
+                className="w-full mt-3 text-gray-500 hover:text-gray-700"
+              >
+                <Plus />
+                Add Task
+              </Button>
+            </DialogTrigger> */}
+            <DialogContent className="w-[95vw] max-w-[425px] mx-auto">
+              <DialogHeader>
+                <DialogTitle>Create New Task</DialogTitle>
+                <p className="text-sm text-gray-600">Add a task to the board</p>
+              </DialogHeader>
+
+              <form className="space-y-4" onSubmit={(e) => {
+                e.preventDefault();
+                const formData = new FormData(e.currentTarget);
+                const taskData: TaskData = {
+                  title: formData.get("title") as string,
+                  description: (formData.get("description") as string) || undefined,
+                  assignee: (formData.get("assignee") as string) || undefined,
+                  dueDate: (formData.get("dueDate") as string) || undefined,
+                  priority: (formData.get("priority") as "low" | "medium" | "high") || "medium",
+                };
+                onCreateTask(taskData);
+              }}>
+                <div className="space-y-2">
+                  <Label>Title *</Label>
+                  <Input
+                    id="title"
+                    name="title"
+                    placeholder="Enter task title"
+                  />
+                </div>
+                <div className="space-y-2">
+                  <Label>Description</Label>
+                  <Textarea
+                    id="description"
+                    name="description"
+                    placeholder="Enter task description"
+                    rows={3}
+                  />
+                </div>
+                <div className="space-y-2">
+                  <Label>Assignee</Label>
+                  <Input
+                    id="assignee"
+                    name="assignee"
+                    placeholder="Who should do this?"
+                  />
+                </div>
+
+                <div className="space-y-2">
+                  <Label>Priority</Label>
+                  <Select name="priority" defaultValue="medium">
+                    <SelectTrigger>
+                      <SelectValue />
+                    </SelectTrigger>
+                    <SelectContent>
+                      {["low", "medium", "high"].map((priority, key) => (
+                        <SelectItem key={key} value={priority}>
+                          {priority.charAt(0).toUpperCase() + priority.slice(1)}
+                        </SelectItem>
+                      ))}
+                    </SelectContent>
+                  </Select>
+                </div>
+
+                <div className="space-y-2">
+                  <Label>Due Date</Label>
+                  <Input type="date" id="dueDate" name="dueDate" />
+                </div>
+
+                <div className="flex justify-end space-x-2 pt-4">
+                  <Button type="submit">Create Task</Button>
+                </div>
+              </form>
+            </DialogContent>
+          </Dialog>
+        </div>
+      </div>
+    </div>
+  );
+}
