+++ components/dashboard/ViewToggle.tsx
@@ -1,0 +1,30 @@
+"use client";
+
+import { Button } from "@/components/ui/button";
+import { Grid3x3, List } from "lucide-react";
+
+interface ViewToggleProps {
+  viewMode: "grid" | "list";
+  onViewModeChange: (mode: "grid" | "list") => void;
+}
+
+export default function ViewToggle({ viewMode, onViewModeChange }: ViewToggleProps) {
+  return (
+    <div className="flex items-center space-x-2 rounded bg-white border p-1">
+      <Button
+        variant={viewMode === "grid" ? "default" : "ghost"}
+        size="sm"
+        onClick={() => onViewModeChange("grid")}
+      >
+        <Grid3x3 />
+      </Button>
+      <Button
+        variant={viewMode === "list" ? "default" : "ghost"}
+        size="sm"
+        onClick={() => onViewModeChange("list")}
+      >
+        <List />
+      </Button>
+    </div>
+  );
+}
